backend/requirements.txt
# ResearchHub AI - Backend Dependencies
# Milestone 1: Activity 1.1 - Create requirements.txt file
# Responsible: Chetan Galphat

fastapi==0.115.5
uvicorn[standard]==0.32.1
groq==0.11.0
numpy==1.26.4
sqlalchemy==2.0.36
alembic==1.14.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.12
pydantic[email]==2.10.3
pydantic-settings==2.6.1
httpx==0.27.0
python-dotenv==1.0.1
scikit-learn==1.5.2
